+++ app/api/tweets/write/route.ts
@@ -1,0 +1,1 @@
+export async function POST(request: Request) {}
