+++ app/tweet/write/write-page.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import Header from "../../../components/header";
+
+export default function WritePage() {
+  return (
+    <div className="text-white">
+      <Header title="Write" />
+    </div>
+  );
+}
